Assignment(28-02-26)/Cloud Inventory Project/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
pymongo==4.6.1
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
flask-cors==4.0.0
cryptography==41.0.7
